requirements.txt
torch
pydantic
fastapi
transformers
bs4
chardet
sentencepiece
lxml
readability-lxml
lxml_html_clean